[hw-1] refactor: Move id checks to service layer

+++ scrapper/src/main/java/backend/academy/scrapper/service/ChatService.java
@@ -1,5 +1,6 @@
 package backend.academy.scrapper.service;
 
+import backend.academy.scrapper.exception.BadRequestException;
 import backend.academy.scrapper.exception.NotFoundException;
 import backend.academy.scrapper.repository.ChatRecord;
 import backend.academy.scrapper.repository.ChatRepository;
@@ -12,10 +13,16 @@ public class ChatService {
     private final ChatRepository chatRepository;
 
     public void register(Long id) {
+        if (id < 0) {
+            throw new BadRequestException(String.format("Невалидный идентификатор чата: %s", id));
+        }
         chatRepository.saveUser(id);
     }
 
     public void unRegister(Long id) {
+        if (id < 0) {
+            throw new BadRequestException(String.format("Невалидный идентификатор чата: %s", id));
+        }
         ChatRecord record = chatRepository.removeUser(id);
         if (record == null) {
             throw new NotFoundException(String.format("Не существует чата с ID: %s", id));
